Addressed comments, added tests
